Move viminfo related functions to a separate file
