+++ src/proto.h
@@ -107,6 +107,9 @@ extern int _stricoll(char *a, char *b);
 # ifdef FEAT_ARABIC
 #  include "arabic.pro"
 # endif
+# ifdef FEAT_VIMINFO
+# include "viminfo.pro"
+# endif
 
 /* These prototypes cannot be produced automatically. */
 int smsg(const char *, ...)
