+++ src/structs.h
@@ -3725,6 +3725,16 @@ typedef enum {
     CDSCOPE_WINDOW	// :lcd
 } cdscope_T;
 
+#if defined(FEAT_VIMINFO) || defined(FEAT_SESSION)
+// Variable flavor
+typedef enum
+{
+    VAR_FLAVOUR_DEFAULT,	/* doesn't start with uppercase */
+    VAR_FLAVOUR_SESSION,	/* starts with uppercase, some lower */
+    VAR_FLAVOUR_VIMINFO		/* all uppercase */
+} var_flavour_T;
+#endif
+
 // argument for mouse_find_win()
 typedef enum {
     IGNORE_POPUP,	// only check non-popup windows
